+++ app/classes/controllers/roles_controller.py
@@ -64,8 +64,8 @@ class RolesController:
             HelperRoles.update_role(role_id, up_data)
 
     @staticmethod
-    def add_role(role_name):
-        return HelperRoles.add_role(role_name)
+    def add_role(role_name, manager):
+        return HelperRoles.add_role(role_name, manager)
 
     class RoleServerJsonType(t.TypedDict):
         server_id: t.Union[str, int]
@@ -92,6 +92,7 @@ class RolesController:
     def add_role_advanced(
         name: str,
         servers: t.Iterable[RoleServerJsonType],
+        manager: int,
     ) -> int:
         """Add a role with a name and a list of servers
 
@@ -102,7 +103,7 @@ class RolesController:
         Returns:
             int: The new role's ID
         """
-        role_id: t.Final[int] = HelperRoles.add_role(name)
+        role_id: t.Final[int] = HelperRoles.add_role(name, manager)
         for server in servers:
             PermissionsServers.get_or_create(
                 role_id, server["server_id"], server["permissions"]
@@ -114,6 +115,7 @@ class RolesController:
         role_id: t.Union[str, int],
         role_name: t.Optional[str],
         servers: t.Optional[t.Iterable[RoleServerJsonType]],
+        manager: int,
     ) -> None:
         """Update a role with a name and a list of servers
 
@@ -152,6 +154,7 @@ class RolesController:
             up_data = {
                 "role_name": role_name,
                 "last_update": Helpers.get_time_as_string(),
+                "manager": manager,
             }
             # TODO: do the last_update on the db side
             HelperRoles.update_role(role_id, up_data)